fix: use platform-adaptive usize for pointer metadata in packed structs (#718)
**Problem**: Using `packed struct(u64)` for pointer fields fails on
32-bit platforms where pointers are 32 bits.

```zig
// BEFORE (fails on WASM32):
pub const PushPointerMetadata = packed struct(u64) { value: *u256 };
pub const JumpStaticMetadata = packed struct(u64) { dispatch: *const anyopaque };

// AFTER (works on all platforms):
pub const PushPointerMetadata = packed struct(usize) { value: *u256 };
pub const JumpStaticMetadata = packed struct(usize) { dispatch: *const anyopaque };
```

**Error on wasm32**:
```
error: backing integer type 'u64' has bit size 64 but the struct fields have a total bit size of 32
```

+++ src/preprocessor/dispatch_metadata.zig
@@ -33,12 +33,12 @@ pub fn DispatchMetadata(comptime FrameType: type) type {
 
         /// Metadata for PUSH operations with values larger than 64 bits.
         /// Contains a pointer to the heap-allocated u256 value.
-        pub const PushPointerMetadata = packed struct(u64) { value: *u256 };
+        pub const PushPointerMetadata = packed struct(usize) { value: *u256 };
 
         /// Metadata for PC opcode containing the program counter value.
         /// Metadata for static jump locations - contains a direct pointer to the jump destination dispatch
         /// This avoids the binary search in findJumpTarget
-        pub const JumpStaticMetadata = packed struct(u64) { dispatch: *const anyopaque };
+        pub const JumpStaticMetadata = packed struct(usize) { dispatch: *const anyopaque };
 
         /// Our instruction set does not match the actual bytecode 1 to 1 if we do fusions so we don't actually track pc during execution
         /// Thus we must provide as metadata to avoid the expensive process of mapping back our instruction index to what pc it came from
